+++ google/src/main/scala/org/broadinstitute/dsde/workbench/google/HttpGoogleIamDAO.scala
@@ -40,7 +40,7 @@ class HttpGoogleIamDAO(appName: String,
                       (implicit system: ActorSystem, executionContext: ExecutionContext)
   extends AbstractHttpGoogleDAO(appName, googleCredentialMode, workbenchMetricBaseName) with GoogleIamDAO {
 
-  @deprecated(message = "This way of instantiating HttpGoogleStorageDAO has been deprecated. Please update to use the primary constructor.", since = "0.14")
+  @deprecated(message = "This way of instantiating HttpGoogleIamDAO has been deprecated. Please update to use the primary constructor.", since = "0.15")
   def this(serviceAccountClientId: String,
            pemFile: String,
            appName: String,
@@ -49,7 +49,7 @@ class HttpGoogleIamDAO(appName: String,
     this(appName, Pem(WorkbenchEmail(serviceAccountClientId), new File(pemFile)), workbenchMetricBaseName)
   }
 
-  @deprecated(message = "This way of instantiating HttpGoogleStorageDAO has been deprecated. Please update to use the primary constructor.", since = "0.14")
+  @deprecated(message = "This way of instantiating HttpGoogleIamDAO has been deprecated. Please update to use the primary constructor.", since = "0.15")
   def this(clientSecrets: GoogleClientSecrets,
            pemFile: String,
            appName: String,
@@ -64,7 +64,7 @@ class HttpGoogleIamDAO(appName: String,
 
   override implicit val service = GoogleInstrumentedService.Iam
 
-  lazy val iam = {
+  private lazy val iam = {
     new Iam.Builder(httpTransport, jsonFactory, googleCredential).setApplicationName(appName).build()
   }